src/main/java/race/MainFX.java
package race;

import javafx.application.Application;
import javafx.geometry.Rectangle2D;
import javafx.scene.image.Image;
import javafx.stage.Screen;
import javafx.stage.Stage;
import race.scene.ModeScene;

import java.io.InputStream;

import static race.picture.PictureReader.readPictureAsInputStream;

public class MainFX extends Application {

    private static final String ICON    = "race/picture/icon.png";
    private static final String TITLE   = "Quadratic Race";

    public static void main(String[] args) {
        launch(args);
    }

    @Override
    public void start(Stage stage) {
        fillHeaderFor(stage);

        stage.setScene(new ModeScene());
        stage.show();

        moveToCenter(stage);
    }

    private void fillHeaderFor(Stage stage) {
        stage.setTitle(TITLE);
        InputStream iconAsInputStream = readPictureAsInputStream(ICON);
        stage.getIcons().add(new Image(iconAsInputStream));
    }

    private void moveToCenter(Stage stage) {
        Rectangle2D primScreenBounds = Screen.getPrimary().getVisualBounds();
        stage.setX((primScreenBounds.getWidth() - stage.getWidth()) / 2);
        stage.setY((primScreenBounds.getHeight() - stage.getHeight()) / 2);
    }
}